requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pdfplumber==0.10.3
spacy==3.7.2
scikit-learn==1.3.2
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.1.3
gunicorn==21.2.0